+++ src/JsonDecoder/Decoder.php
@@ -5,11 +5,11 @@ namespace JsonMachine\JsonDecoder;
 @trigger_error(sprintf(
     'Interface %s is deprecated. Use %s instead.',
     Decoder::class,
-    ChunkDecoder::class
+    ItemDecoder::class
 ), E_USER_DEPRECATED);
 
 /**
- * @deprecated Use ChunkDecoder instead.
+ * @deprecated Use ItemDecoder instead.
  */
 interface Decoder
 {